pyproject.toml
[tool.poetry]
name = "tempeture-generator"
version = "0.1.0"
description = "Genearator of data tempeture from sensor in area"
authors = ["Artem Eroshev"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Flask-APScheduler = "^1.12.3"
kafka-python = "^2.0.2"
Faker = "^13.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
